Update strings.xml
Thank you, you are a very wonderful person, it is nice that you are going to meet everyone. If possible, add information about the person who made the translation, something like: "The translation was made by Mykola Popovych for my beloved Victoria."

I also want free help with the translation of all tabs such as color, effects, segments, presets, and all other inscriptions in the application, because when I enter the device, everything is in English, or do I need to contact the seller of the LED lamp?